+++ processing/src/test/java/io/druid/query/aggregation/LongMaxAggregationTest.java
@@ -65,11 +65,11 @@ public class LongMaxAggregationTest
     aggregate(selector, agg);
 
     Assert.assertEquals(values[2], ((Long) agg.get()).longValue());
-    Assert.assertEquals(values[2], agg.getLong());
+    Assert.assertEquals(values[2], agg.getLong().longValue());
     Assert.assertEquals((float) values[2], agg.getFloat(), 0.0001);
 
     agg.reset();
-    Assert.assertEquals(Long.MIN_VALUE, agg.getLong());
+    Assert.assertEquals(Long.MIN_VALUE, agg.getLong().longValue());
   }
 
   @Test
@@ -86,7 +86,7 @@ public class LongMaxAggregationTest
     aggregate(selector, agg, buffer, 0);
 
     Assert.assertEquals(values[2], ((Long) agg.get(buffer, 0)).longValue());
-    Assert.assertEquals(values[2], agg.getLong(buffer, 0));
+    Assert.assertEquals(values[2], agg.getLong(buffer, 0).longValue());
     Assert.assertEquals((float)values[2], agg.getFloat(buffer, 0), 0.0001);
   }
